More order summary styling
Styling the container for the order summary and the elements inside it. Like I said, the goal for this is so the order summary is on the top and on the right side of the page. I have to fix my js so that this new code works with it